requirements.txt
Flask~=3.0.0
Flask-Limiter~=3.5.0
PyYAML~=6.0.1
